requirements.txt
certifi==2017.7.27.1
emoji==0.4.5
future==0.16.0
peewee==2.10.2
psycopg2==2.7.3.1
python-telegram-bot==8.0
